requirements.txt
matplotlib
numpy
pypng
bottle